\name{bland.altman.PEFR.1986}
\alias{bland.altman.PEFR.1986}
\docType{data}
\title{Sample PEFR comparison data from Bland-Altman (1986)}
\description{This is the sample PEFR data set from the 1986 Lancet paper written by Bland and Altman. I do not claim any copyright on the data - this is meant to allow testing of the function. I encourage future package authors to use the .rda file if they so wish.}
\usage{data("bland.altman.PEFR.1986")}
\format{
  A data frame with 17 observations on the following 4 variables.
  \describe{
    \item{\code{WrightFirst}}{a numeric vector}
    \item{\code{WrightSecond}}{a numeric vector}
    \item{\code{MiniWrightFirst}}{a numeric vector}
    \item{\code{MiniWrightSecond}}{a numeric vector}
  }
}
\references{Bland, J. M., & Altman, D. (1986). Statistical methods for assessing agreement between two methods of clinical measurement. The Lancet, 327(8476), 307-310. PMID:2868172. http://dx.doi.org/10.1016/S0140-6736(86)90837-8}
\examples{
data(bland.altman.PEFR.1986)}
\keyword{datasets}
