% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.dataset.fibre.r
\name{blandr.dataset.fibre}
\alias{blandr.dataset.fibre}
\title{Function to load D'arbela mean velocity of circumferential fibre shortening dataset from internet}
\usage{
blandr.dataset.fibre()
}
\value{
converted.from.dct A data frame containing the dataset
}
\description{
Loads the D'arbela mean velocity of circumferential fibre shortening dataset from Martin Bland's website.
}
\note{
The function converts the STATA DCT data format into a data frame that R can process.
}
\examples{
blandr.dataset.fibre()
pefr.data <- blandr.dataset.fibre()

}
\references{
https://www-users.york.ac.uk/~mb55/datasets/pefr.dct - The D'arbela mean velocity of circumferential fibre shortening dataset from Martin Bland's website

https://www-users.york.ac.uk/~mb55/datasets/datasets.htm - Martin Bland's example data-set webpage

Bland JM, Altman DG. (1986) Statistical methods for assessing agreement between two methods of clinical measurement. Lancet i, 307-310.
}
\author{
Deepankar Datta <deepankar.datta@gmail.com>
}
