% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.dataset.o2sats.r
\name{blandr.dataset.o2sats}
\alias{blandr.dataset.o2sats}
\title{Function to load Bland-Altman oxygen saturation dataset from internet}
\usage{
blandr.dataset.o2sats()
}
\value{
converted.from.dct A data frame containing the dataset
}
\description{
Loads the Bland-Altman oxygen saturation dataset from Martin Bland's website.
}
\note{
The function converts the STATA DCT data format into a data frame that R can process.
}
\examples{
blandr.dataset.o2sats()
pefr.data <- blandr.dataset.o2sats()

}
\references{
https://www-users.york.ac.uk/~mb55/datasets/sealey.dct - The oxygen saturation dataset from Martin Bland's website

https://www-users.york.ac.uk/~mb55/datasets/datasets.htm - Martin Bland's example data-set webpage

Bland JM, Altman DG. (1986) Statistical methods for assessing agreement between two methods of clinical measurement. Lancet i, 307-310.
}
\author{
Deepankar Datta <deepankar.datta@gmail.com>
}
