% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.dataset.pefr.r
\name{blandr.dataset.pefr}
\alias{blandr.dataset.pefr}
\title{Function to load Bland-Altman PEFR dataset from internet}
\usage{
blandr.dataset.pefr()
}
\value{
converted.from.dct A data frame containing the dataset
}
\description{
Loads the Bland-Altman PEFR dataset from Martin Bland's website.
}
\note{
The function converts the STATA DCT data format into a data frame that R can process.
}
\examples{
blandr.dataset.pefr()
pefr.data <- blandr.dataset.pefr()

}
\references{
https://www-users.york.ac.uk/~mb55/datasets/pefr.dct - The PEFR dataset from Martin Bland's website

https://www-users.york.ac.uk/~mb55/datasets/datasets.htm - Martin Bland's example data-set webpage

Bland JM, Altman DG. (1986) Statistical methods for assessing agreement between two methods of clinical measurement. Lancet i, 307-310.
}
\author{
Deepankar Datta \href{mailto:deepankar.datta@gmail.com}{deepankar.datta@gmail.com}
}
