% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.dataset.sbp.r
\name{blandr.dataset.sbp}
\alias{blandr.dataset.sbp}
\title{Function to load Close systolic blood pressure dataset from internet}
\usage{
blandr.dataset.sbp()
}
\value{
converted.from.dct A data frame containing the dataset
}
\description{
Loads the Close systolic blood pressure dataset from Martin Bland's website.
}
\note{
The function converts the STATA DCT data format into a data frame that R can process.
}
\examples{
blandr.dataset.sbp()
pefr.data <- blandr.dataset.sbp()

}
\references{
https://www-users.york.ac.uk/~mb55/datasets/pefr.dct - The Close systolic blood pressure dataset from Martin Bland's website

https://www-users.york.ac.uk/~mb55/datasets/datasets.htm - Martin Bland's example data-set webpage

Bland JM, Altman DG. (1995) Comparing methods of measurement: why plotting difference against standard method is misleading. Lancet, 346, 1085-7.
}
\author{
Deepankar Datta <deepankar.datta@gmail.com>
}
