% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.display.and.draw.r
\name{blandr.display.and.draw}
\alias{blandr.display.and.draw}
\title{(DEPRECATED) Bland-Altman display and draw for R}
\usage{
blandr.display.and.draw(method1, method2, plotter = "ggplot",
  method1name = "Method 1", method2name = "Method 2",
  plotTitle = "Bland-Altman plot for comparison of 2 methods",
  sig.level = 0.95, annotate = FALSE, ciDisplay = TRUE,
  ciShading = FALSE, normalLow = FALSE, normalHigh = FALSE,
  lowest_y_axis = FALSE, highest_y_axis = FALSE, point_size = 0.8)
}
\description{
(DEPRECATED) Stub function: calls both the display and plots functions (in that order).
Uses the same parameters as the plot and display functions to allow easy all-in-one use.
This function may be deprecated in future, as you really can use the functions easily
separately.
}
\examples{
# Generates two random measurements
measurement1 <- rnorm(100)
measurement2 <- rnorm(100)

# Generates a plot, with no optional arguments
blandr.display.and.draw( measurement1 , measurement2 )

# Generates a plot, with title
blandr.display.and.draw( measurement1 , measurement2 , plotTitle = 'Bland-Altman example plot' )

}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
