% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.output.report.r
\name{blandr.output.report}
\alias{blandr.output.report}
\title{Bland-Altman report generating function (BETA)}
\usage{
blandr.output.report(method1, method2)
}
\arguments{
\item{method1}{A list of numbers for the first method}

\item{method2}{A list of numbers for the second method}
}
\description{
Generates a report for the Bland-Altman statistics. The function passes data to a R markdown
}
\examples{
# Generates two random measurements
measurement1 <- rnorm(100)
measurement2 <- rnorm(100)

blandr.output.report( measurement1 , measurement2 )

}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
