% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.output.text.r
\name{blandr.output.text}
\alias{blandr.output.text}
\title{(DEPRECATED) Bland-Altman summary statistics display function}
\usage{
blandr.output.text(method1, method2, sig.level = 0.95)
}
\arguments{
\item{method1}{A list of numbers.}

\item{method2}{A list of numbers.}

\item{sig.level}{(Optional) Two-tailed significance level. Expressed from 0 to 1. Defaults to 0.95.}
}
\description{
(DEPRECATED) Displays results of Bland-Altman analysis in a nicer text format. Relies on the blandr.statistics function in the package.
}
\examples{
# Generates two random measurements
measurement1 <- rnorm(100)
measurement2 <- rnorm(100)

# Displays basic statistics for the two measurements in a readable form
blandr.output.text( measurement1 , measurement2 )

}
\author{
Deepankar Datta <deepankar.datta@gmail.com>
}
