% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.plot.normality.r
\name{blandr.plot.normality}
\alias{blandr.plot.normality}
\title{Bland-Altman histogram and density plot}
\usage{
blandr.plot.normality(statistics.results)
}
\arguments{
\item{statistics.results}{A list of statistics generated by the blandr.statistics function: see the function's return list to see what variables are passed to this function}
}
\description{
Generates  a combined histogram and density curve for Bland-Altman differences
}
\author{
Deepankar Datta <deepankar.datta@gmail.com>
}
