% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.plot.qq.r
\name{blandr.plot.qq}
\alias{blandr.plot.qq}
\title{Bland-Altman differences QQ plot}
\usage{
blandr.plot.qq(statistics.results)
}
\arguments{
\item{statistics.results}{A list of statistics generated by the blandr.statistics function: see the function's return list to see what variables are passed to this function}
}
\description{
Generates a QQ plot for Bland-Altman differences
}
\author{
Deepankar Datta \href{mailto:deepankar.datta@gmail.com}{deepankar.datta@gmail.com}
}
