% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.plot.r
\name{blandr.plot}
\alias{blandr.plot}
\title{(DEPRECATED) Bland-Altman drawing function for R}
\usage{
blandr.plot(method1, method2, plotter = "ggplot", method1name = "Method 1",
  method2name = "Method 2",
  plotTitle = "Bland-Altman plot for comparison of 2 methods",
  sig.level = 0.95, annotate = FALSE, ciDisplay = TRUE,
  ciShading = TRUE, normalLow = FALSE, normalHigh = FALSE,
  lowest_y_axis = FALSE, highest_y_axis = FALSE, point_size = 0.8)
}
\description{
(DEPRECATED) Re-directs to blandr.draw.r
}
