% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandr.statistics.r
\name{blandr.statistics.formula}
\alias{blandr.statistics.formula}
\title{@rdname blandr.statistics}
\usage{
\method{blandr.statistics}{formula}(formula, data = parent.frame(), ...)
}
\description{
@rdname blandr.statistics
}
