% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.load.R
\name{dataset.load}
\alias{dataset.load}
\title{Function to load example data sets}
\usage{
dataset.load(dataset.name)
}
\arguments{
\item{dataset.name}{Loads the requisite data set. See the description for further details.}
}
\value{
example.dataset A data frame containing the requisite dataset
}
\description{
Loads example data sets from the internet.
}
\note{
Dataset 1 ("1","PEFR","pefr") - Bland Altman PEFR dataset (from \code{dataset.pefr})

Dataset 2 ("2","o2sats","sealey") - Selaey oxygen saturations dataset (from \code{dataset.o2sats})

Dataset 3 ("3","fibre","darbela") - D'arbela mean velocity of circumferential fibre shortening dataset (from \code{dataset.fibre})

Dataset 4 ("4","sbp","close") - Close systolic blood pressure dataset (from \code{dataset.sbp})
}
\examples{
dataset.load( "pefr" )
pefr.data <- dataset.load( "pefr" )

}
\author{
Deepankar Datta <deepankardatta@nhs.net>
}
