% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.pefr.R
\name{dataset.pefr}
\alias{dataset.pefr}
\title{Function to load Bland-Altman PEFR dataset from internet}
\usage{
dataset.pefr()

dataset.pefr()
}
\value{
converted.from.dct A data frame containing the dataset

converted.from.dct A data frame containing the dataset
}
\description{
Loads the Bland-Altman PEFR dataset from Martin Bland's website.

Loads the Bland-Altman PEFR dataset from Martin Bland's website.
}
\note{
The function converts the STATA DCT data format into a data frame that R can process.

The function converts the STATA DCT data format into a data frame that R can process.
}
\examples{
dataset.pefr()
pefr.data <- dataset.pefr()

dataset.pefr()
pefr.data <- dataset.pefr()

}
\references{
https://www-users.york.ac.uk/~mb55/datasets/pefr.dct - The PEFR dataset from Martin Bland's website

https://www-users.york.ac.uk/~mb55/datasets/datasets.htm - Martin Bland's example data-set webpage

Bland JM, Altman DG. (1986) Statistical methods for assessing agreement between two methods of clinical measurement. Lancet i, 307-310.

https://www-users.york.ac.uk/~mb55/datasets/pefr.dct - The PEFR dataset from Martin Bland's website

https://www-users.york.ac.uk/~mb55/datasets/datasets.htm - Martin Bland's example data-set webpage

Bland JM, Altman DG. (1986) Statistical methods for assessing agreement between two methods of clinical measurement. Lancet i, 307-310.
}
\author{
Deepankar Datta <deepankardatta@nhs.net>

Deepankar Datta <deepankardatta@nhs.net>
}
