% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovibanormality.h.R
\name{jamoviBAnormality}
\alias{jamoviBAnormality}
\title{Bland-Altman normality testing}
\usage{
jamoviBAnormality(data, method1, method2)
}
\arguments{
\item{data}{.}

\item{method1}{.}

\item{method2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$plotNormality} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Bland-Altman normality testing
}
