% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovibaplothistogram.h.R
\name{jamoviBAplotHistogram}
\alias{jamoviBAplotHistogram}
\title{Bland-Altman differences histogram and density curve}
\usage{
jamoviBAplotHistogram(data, method1, method2)
}
\arguments{
\item{data}{.}

\item{method1}{.}

\item{method2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$plotNormality} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Bland-Altman differences histogram and density curve
}
