% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovibaplotqq.h.R
\name{jamoviBAplotQQ}
\alias{jamoviBAplotQQ}
\title{Differences QQ plot}
\usage{
jamoviBAplotQQ(data, method1, method2)
}
\arguments{
\item{data}{.}

\item{method1}{.}

\item{method2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Differences QQ plot
}
