% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovibastats.h.R
\name{jamoviBAstats}
\alias{jamoviBAstats}
\title{Bland-Altman Raw Statistics}
\usage{
jamoviBAstats(data, method1, method2)
}
\arguments{
\item{data}{.}

\item{method1}{.}

\item{method2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
}
}
\description{
Bland-Altman Raw Statistics
}
