% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.blandr.R
\name{plot.blandr}
\alias{plot.blandr}
\title{Plot method for \code{blandr} objects}
\usage{
\method{plot}{blandr}(x, method1name = "Method 1",
  method2name = "Method 2",
  plotTitle = "Bland-Altman plot for comparison of 2 methods",
  sig.level = 0.95, LoA.mode = 1, annotate = FALSE,
  ciDisplay = TRUE, ciShading = TRUE, normalLow = FALSE,
  normalHigh = FALSE, lowest_y_axis = FALSE, highest_y_axis = FALSE,
  point_size = 0.8, ...)
}
\arguments{
\item{x}{a blandr class object, the results from \code{blandr.statistics}}

\item{method1name}{(Optional) Plotting name for 1st method, default 'Method 1'}

\item{method2name}{(Optional) Plotting name for 2nd method, default 'Method 2'}

\item{plotTitle}{(Optional) Title name, default 'Bland-Altman plot for comparison of 2 methods'}

\item{annotate}{(Optional) TRUE/FALSE switch to provides annotations to plot, default=FALSE}

\item{ciDisplay}{(Optional) TRUE/FALSE switch to plot confidence intervals for bias and limits of agreement, default=TRUE}

\item{ciShading}{(Optional) TRUE/FALSE switch to plot confidence interval shading to plot, default=TRUE}

\item{normalLow}{(Optional) If there is a normal range, entering a continuous variable will plot a vertical line on the plot to indicate its lower boundary}

\item{normalHigh}{(Optional) If there is a normal range, entering a continuous variable will plot a vertical line on the plot to indicate its higher boundary}

\item{point_size}{(Optional) Size of marker for each dot. Default is cex=0.8}

\item{...}{other arguments}

\item{plot.limits}{A list of statistics generated by the blandr.plot.limits function to define the extent of the x- and y- axes: see the function's return list to see what variables are passed to this function}
}
\description{
Plot a Bland-Altman chart from the results of \code{blandr} objects
using the in-built rplot drawing functions
}
\examples{

results = blandr.statistics( Method.B ~ Method.A , data = giavarina.2015 )
results
plot(results)

}
