% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.blandr.R
\name{print.blandr}
\alias{print.blandr}
\title{Print method for \code{blandr} objects}
\usage{
\method{print}{blandr}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{the results from \code{statistics}.}

\item{digits}{number of significant digits to be used.}

\item{...}{other arguments.}
}
\description{
Compactly print the statistics used to construct a Bland-Altman plot
}
\examples{

results = statistics(Method.B ~ Method.A, data = giavarina.2015)
results
print(results)

}
