% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.blandr.R
\name{summary.blandr}
\alias{summary.blandr}
\title{Summary method for \code{blandr} objects}
\usage{
\method{summary}{blandr}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{the results from \code{blandr.statistics}.}

\item{digits}{number of significant digits to be used.}

\item{...}{other arguments.}
}
\description{
Displays a summary of a blandr object, the result from \code{blandr.statistics}
}
\examples{

results = blandr.statistics(Method.B ~ Method.A, data = giavarina.2015)
summary(results)

}
