% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamovibaplothistogram.h.R
\name{jamoviBAplotHistogram}
\alias{jamoviBAplotHistogram}
\title{Differences Histogram and Density Curve}
\usage{
jamoviBAplotHistogram(data, method1, method2)
}
\arguments{
\item{data}{.}

\item{method1}{.}

\item{method2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
\code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Differences Histogram and Density Curve
}
