% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.blandr.R
\name{plot.blandr}
\alias{plot.blandr}
\title{Plot method for \code{blandr} objects}
\usage{
\method{plot}{blandr}(x, method1name = "Method 1",
  method2name = "Method 2",
  plotTitle = "Bland-Altman plot for comparison of 2 methods",
  sig.level = 0.95, LoA.mode = 1, annotate = FALSE,
  ciDisplay = TRUE, ciShading = TRUE, normalLow = FALSE,
  normalHigh = FALSE, lowest_y_axis = FALSE, highest_y_axis = FALSE,
  point_size = 0.8, overlapping = FALSE, plotter = "ggplot",
  x.plot.mode = "means", y.plot.mode = "difference",
  plotProportionalBias = FALSE, plotProportionalBias.se = TRUE,
  assume.differences.are.normal = TRUE, ...)
}
\arguments{
\item{x}{a blandr class object, the results from \code{blandr.statistics}.}

\item{...}{other arguments.}
}
\description{
Plot a Bland-Altman chart from the results of \code{blandr} objects
using the in-built rplot drawing functions
}
\examples{

results = blandr.statistics( Method.B ~ Method.A , data = giavarina.2015 )
results
plot(results)

}
