% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.blandr.R
\name{print.blandr}
\alias{print.blandr}
\title{Print method for \code{blandr} objects}
\usage{
\method{print}{blandr}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{the results from \code{blandr.statistics}.}

\item{digits}{number of significant digits to be used.}

\item{...}{other arguments.}
}
\description{
Compactly print the statistics used to construct a Bland-Altman plot
}
\examples{

results = blandr.statistics.formula(Method.B ~ Method.A, data = giavarina.2015)
results

}
\author{
Deepankar Datta \href{mailto:deepankar.datta@gmail.com}{deepankar.datta@gmail.com}
}
